% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthbasedreconstructPOFD.R
\name{envelope}
\alias{envelope}
\title{Envelope algorithm}
\usage{
envelope(data, i, max_iter = 10)
}
\arguments{
\item{data}{Data matrix.}

\item{i}{Index of curve.}

\item{max_iter}{Maximum number of nearest curves considered in for loop. By
default, max_iter = 5.}
}
\value{
Envelope (set of indices).
}
\description{
Code for obtaining the envelope Ji of a curve specified by the index i. The
implementation is based on Algorithm 1 in [1]. References:
}
\details{
[1] Elías, A., Jiménez, R., & Shang, H. L. (2023). Depth-based reconstruction
method for incomplete functional data. Computational Statistics, 38(3),
1507-1535.
}
