% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{exampleData}
\format{
A list with three data sets (functions by columns):
\describe{
\item{PoFDintervals}{Partially observed functional data in intervals}
\item{PoFDextremes}{Partially Observed functional data with missing intervals at the extremes}
\item{PoFDextremes_outliers}{Same as above but including two magnitude and shape outliers}
}
}
\usage{
exampleData
}
\description{
An illustrative Functional Gaussian processes with different partially observed patterns
with outliers and without outliers.
}
\examples{
data(exampleData)
plot_PoFD(exampleData$PoFDintervals)

}
\references{
Elías, Antonio, Jiménez, Raúl, Paganoni, Anna M. and Sangalli, Laura M. (2020). Integrated Depths for Partially ObservedFunctional Data.
}
\keyword{datasets}
