% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthbasedreconstructPOFD.R
\name{learningK}
\alias{learningK}
\title{Learning k}
\usage{
learningK(data, i, J)
}
\arguments{
\item{data}{Data matrix.}

\item{i}{Index of curve.}

\item{J}{Envelope.}
}
\value{
Optimal parameter k.
}
\description{
This function chooses an optimal parameter k (denoting the number of curves
within the envelope that are considered for reconstruction).
}
