% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPoFD.R
\name{plot_PoFD}
\alias{plot_PoFD}
\title{Plot Partially Observed Functional Data}
\usage{
plot_PoFD(data)
}
\arguments{
\item{data}{matrix p by n, being n the number of functions and p the number of grid points.}
}
\value{
Plot of the partially observed functional data and the proportion of observed functions
at each time point.
}
\description{
Plot the sample of partially observed curves and the proportion of observed functions.
}
\examples{
#plot_PoFD(data)

}
