% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliergramPOFD.R
\name{outliergramPOFD}
\alias{outliergramPOFD}
\title{Outliergram for Partially Observed Functional Data}
\usage{
outliergramPOFD(data, fshape = 1.5, p1 = 1, p2 = 0)
}
\arguments{
\item{data}{matrix p by n, being n the number of functions and p the number of grid points.}

\item{fshape}{inflation of the outliergram that determine the shape outlier rule.}

\item{p1}{parameter of the outliergram for resampling method. Default = 1.}

\item{p2}{parameter of the outliergram for resampling method. Default = 0.}
}
\value{
a list with the functional outliergram for PoDF and the shape outliers.
}
\description{
Plots the Outliergram for PoFD and returns the shape outliers.
}
\examples{
data(exampleData)
outliergramPOFD(exampleData$PoFDextremes_outliers, fshape = 1.5, p1 = 1, p2 = 0)

}
\references{
Arribas-Gil, A. and Romo, J. (2014).  Shape outlier detection and visualization for functional data:  the outliergram.Biostatistics, 15(4):603–619.
}
